separate beam layout from voice element
