+++ src/write/abc_voice_element.js
@@ -69,29 +69,6 @@ VoiceElement.prototype.setLimit = function(member, child) {
 		this.specialY[member] = Math.max(this.specialY[member], specialY[member]);
 };
 
-VoiceElement.prototype.moveDecorations = function(beam) {
-	var padding = 1.5; // This is the vertical padding between elements, in pitches.
-	for (var ch = 0; ch < beam.elems.length; ch++) {
-		var child = beam.elems[ch];
-		if (child.top) {
-			// We now know where the ornaments should have been placed, so move them if they would overlap.
-			var top = beam.yAtNote(child);
-			for (var i = 0; i < child.children.length; i++) {
-				var el = child.children[i];
-				if (el.klass === 'ornament') {
-					if (el.bottom - padding < top) {
-						var distance = top - el.bottom + padding; // Find the distance that it needs to move and add a little margin so the element doesn't touch the beam.
-						el.bottom += distance;
-						el.top += distance;
-						el.pitch += distance;
-						top = child.top = el.top;
-					}
-				}
-			}
-		}
-	}
-};
-
 VoiceElement.prototype.adjustRange = function(child) {
 	if (child.bottom !== undefined)
 		this.bottom = Math.min(this.bottom, child.bottom);
@@ -216,29 +193,6 @@ VoiceElement.prototype.shiftRight = function (dx) {
 	child.setX(child.x+dx);
 	this.minx+=dx;
 	this.nextx+=dx;
-};
-
-VoiceElement.prototype.layoutBeams = function() {
-	for (var i = 0; i < this.beams.length; i++) {
-		if (this.beams[i].layout) {
-			this.beams[i].layout();
-			this.moveDecorations(this.beams[i]);
-			// The above will change the top and bottom of the abselem children, so see if we need to expand our range.
-			for (var j = 0; j < this.beams[i].elems.length; j++) {
-				this.adjustRange(this.beams[i].elems[j]);
-			}
-		}
-	}
-	// Now we can layout the triplets
-	for (i = 0; i < this.otherchildren.length; i++) {
-		var child = this.otherchildren[i];
-		if (child.layout) {
-			child.layout();
-			this.adjustRange(child);
-		}
-	}
-	this.staff.top = Math.max(this.staff.top, this.top);
-	this.staff.bottom = Math.min(this.staff.bottom, this.bottom);
 };
 
 module.exports = VoiceElement;
