+++ src/write/layout/beams.js
@@ -1,0 +1,47 @@
+var layoutBeams = function(voice) {
+	for (var i = 0; i < voice.beams.length; i++) {
+		if (voice.beams[i].layout) {
+			voice.beams[i].layout();
+			moveDecorations(voice.beams[i]);
+			// The above will change the top and bottom of the abselem children, so see if we need to expand our range.
+			for (var j = 0; j < voice.beams[i].elems.length; j++) {
+				voice.adjustRange(voice.beams[i].elems[j]);
+			}
+		}
+	}
+	// Now we can layout the triplets
+	for (i = 0; i < voice.otherchildren.length; i++) {
+		var child = voice.otherchildren[i];
+		if (child.layout) {
+			child.layout();
+			voice.adjustRange(child);
+		}
+	}
+	voice.staff.top = Math.max(voice.staff.top, voice.top);
+	voice.staff.bottom = Math.min(voice.staff.bottom, voice.bottom);
+};
+
+function moveDecorations(beam) {
+	var padding = 1.5; // This is the vertical padding between elements, in pitches.
+	for (var ch = 0; ch < beam.elems.length; ch++) {
+		var child = beam.elems[ch];
+		if (child.top) {
+			// We now know where the ornaments should have been placed, so move them if they would overlap.
+			var top = beam.yAtNote(child);
+			for (var i = 0; i < child.children.length; i++) {
+				var el = child.children[i];
+				if (el.klass === 'ornament') {
+					if (el.bottom - padding < top) {
+						var distance = top - el.bottom + padding; // Find the distance that it needs to move and add a little margin so the element doesn't touch the beam.
+						el.bottom += distance;
+						el.top += distance;
+						el.pitch += distance;
+						top = child.top = el.top;
+					}
+				}
+			}
+		}
+	}
+};
+
+module.exports = layoutBeams;
